+++ CHANGELOG.md
@@ -9,10 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 - Error adding actions to questions. ([#144][i144])
 - Rewards link is broken on Admin Sidebar. ([#193][i193])
 - Missing translation: admin/question/model.published_at. ([#192][i192])
+- Users can't change their password. ([#135][i135])
 
 [i144]: https://github.com/pacoorozco/gamify-laravel/issues/144
 [i193]: https://github.com/pacoorozco/gamify-laravel/issues/193
 [i192]: https://github.com/pacoorozco/gamify-laravel/issues/192
+[i135]: https://github.com/pacoorozco/gamify-laravel/issues/135
 
 ## 2.6.0 - 2021-02-22